Include jquery.flot.stack* for using in multiple panels
When using multiple panels that use the stacked graphs
(e.g. histogram and multifieldhistogram) we need to inform
grunt-requirejs plugin that jquery.flot.stack* is one the
modules to prevent double-inclusion and then problems
in the histogram rendering.

See https://github.com/elasticsearch/kibana/pull/1296

NOTE: this change is needed for getting pull#1296 to work
properly but can be merged indipendently from it.